Safari-611.1.10/LayoutTests/fast/frames/sandbox-attribute-expected.txt

Checks that HTMLIframeElement.sandbox is a DOMTokenList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.sandbox.__proto__ is DOMTokenList.prototype

iframe.sandbox = 'allow-popups   allow-scripts  allow-scripts  allow-same-origin'
PASS iframe.sandbox.length is 3
PASS iframe.sandbox[0] is "allow-popups"
PASS iframe.sandbox[1] is "allow-scripts"
PASS iframe.sandbox[2] is "allow-same-origin"
PASS String(iframe.sandbox) is "allow-popups   allow-scripts  allow-scripts  allow-same-origin"
PASS iframe.getAttribute('sandbox') is "allow-popups   allow-scripts  allow-scripts  allow-same-origin"

iframe.setAttribute('sandbox', 'allow-popups allow-scripts')
PASS iframe.sandbox.length is 2
PASS iframe.sandbox[0] is "allow-popups"
PASS iframe.sandbox[1] is "allow-scripts"
PASS String(iframe.sandbox) is "allow-popups allow-scripts"
PASS iframe.getAttribute('sandbox') is "allow-popups allow-scripts"
PASS successfullyParsed is true

TEST COMPLETE
